+++ ir.h
@@ -13,8 +13,8 @@ typedef std::experimental::optional<std::string> optstr_t;
 typedef std::vector<Temp> TempList;
 
 struct Temp {
-    int32_t stackOffset = -1;
-	Temp(int32_t offset) : stackOffset(offset) {}
+    int64_t stackOffset = -1;
+	Temp(int64_t offset) : stackOffset(offset) {}
 };
 
 enum class IrOp {
